Store colours as bgra for easier conversion from juce::Colour

+++ src/fontstash.h
@@ -1008,7 +1008,7 @@ static FONSglyph* fons__getGlyph(FONScontext* stash, int fontid, unsigned int co
         if (font->notDef < 0) {  // notdef not created yet
             FONSglyph* notdefGlyph = fons__getGlyph(stash, fontid, 0xFFFF, flags);
             if (!notdefGlyph) { glyph = NULL; goto done; }  // this can happen if atlas is full
-            font->notDef = notdefGlyph - font->glyphs;
+            font->notDef = (int)(notdefGlyph - font->glyphs);
             assert(font->notDef >= 0 && font->notDef < font->nglyphs);
         }
         g = -1;  // this glyph will reference notDef
